build(cmake)!: modernize build system and project structure (#17)
* build(cmake)!: modernize build system and project structure

* fix(data): remove invalid default constructor

* fix(data): remove unused numLine variable

* fix(data): omit unused rawData parameter name in synchronize

* fix(performance): resolve implicit int to float conversion

* fix(portfolio): resolve compiler warnings mapped to errors

* fix(utils): resolve implicit sign conversion warning

* fix(strategy): resolve smacrossover compiler warnings

* test(portfolio): decouple tests from datahandler

* fix(benchmark): resolve 64-to-32 bit integer precision loss

* ci: update github actions for modern cmake presets

* fix(data): rename symDict parameter

* fix(smacrossover): resolve shadow warning in constructor

* fix(benchmarks): explicitly cast int64_t to double in dummy CSV generation

+++ benchmarks/bm_data.cpp
@@ -9,7 +9,7 @@
 #include "backtest-cpp/types.h"
 #include "backtest-cpp/utils.h"
 
-static void CreateDummyCSV(const std::string& filepath, int numRows) {
+static void CreateDummyCSV(const std::string& filepath, int64_t numRows) {
     std::ofstream file(filepath);
     file << "timestamp,open,high,low,close,volume\n";
 
@@ -18,8 +18,8 @@ static void CreateDummyCSV(const std::string& filepath, int numRows) {
 
     const int64_t hour_in_ns = 3600000000000LL;
 
-    for (int i = 0; i < numRows; i++) {
-        double basePrice = 3700.0 + (i % 100);
+    for (int64_t i = 0; i < numRows; i++) {
+        double basePrice = 3700.0 + static_cast<double>(i % 100);
 
         // Write the 64-bit nanosecond integer to the file
         file << (base_timestamp_ns + i * hour_in_ns) << "," << basePrice << ","
@@ -30,7 +30,7 @@ static void CreateDummyCSV(const std::string& filepath, int numRows) {
 
 // Benchmark the CSV loading process
 static void BM_DataHandler_LoadCSV(benchmark::State& state) {
-    const int numRows = state.range(0);
+    const int64_t numRows = state.range(0);
     const std::string filepath = "bm_temp_load_" + std::to_string(numRows) + ".csv";
 
     CreateDummyCSV(filepath, numRows);
@@ -55,7 +55,7 @@ static void BM_DataHandler_LoadCSV(benchmark::State& state) {
 
 // Benchmark retrieving bars sequentially
 static void BM_DataHandler_GetNextBars(benchmark::State& state) {
-    const int numRows = state.range(0);
+    const int64_t numRows = state.range(0);
     const std::string filepath = "bm_temp_get_" + std::to_string(numRows) + ".csv";
     CreateDummyCSV(filepath, numRows);
